seekto: add a bunch of comments and cleanups
